Se implementa JWT para el login guardar el token de manera local

+++ src/main/java/com/cursojava/curso/controllers/UsuarioController.java
@@ -2,6 +2,7 @@ package com.cursojava.curso.controllers;
 
 import com.cursojava.curso.dao.UsuarioDao;
 import com.cursojava.curso.models.Usuario;
+import com.cursojava.curso.utils.JWTUtil;
 import de.mkammerer.argon2.Argon2;
 import de.mkammerer.argon2.Argon2Factory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,16 +16,20 @@ public class UsuarioController {
 
     @Autowired
     private UsuarioDao usuarioDao;
+    @Autowired
+    private JWTUtil jwtUtil;
 
     @RequestMapping(value = "api/usuario/{id}", method = RequestMethod.GET)
     public Usuario getUsuario(@PathVariable Long id) {
-        Usuario usuario = new Usuario(id, "Jeferson", "Jimenez",
-                "jeferson.jl99@gmail.com", "3228186615", "Holaquease");
+        Usuario usuario = new Usuario(id, "Jeferson", "Jimenez", "jeferson.jl99@gmail.com", "3228186615", "Holaquease");
         return usuario;
     }
 
     @RequestMapping(value = "api/usuario/{id}", method = RequestMethod.DELETE)
-    public void deleteUsuario(@PathVariable Long id) {
+    public void deleteUsuario(@RequestHeader(value = "Authorization") String token, @PathVariable Long id) {
+        if (validarToken(token)) {
+            return;
+        }
         usuarioDao.delete(id);
     }
 
@@ -38,9 +43,18 @@ public class UsuarioController {
     }
 
     @RequestMapping(value = "api/usuarios", method = RequestMethod.GET)
-    public List<Usuario> getUsuarios() {
+    public List<Usuario> getUsuarios(@RequestHeader(value = "Authorization") String token) {
+        if (validarToken(token)) {
+            return null;
+        }
         return usuarioDao.getUsuarios();
     }
+
+    public boolean validarToken(String token) {
+        String uduarioId = jwtUtil.getKey(token);
+        return uduarioId != null;
+    }
+
 //    @RequestMapping(value = "usuarios")
 //    public List<Usuario> getUsuarios() {
 //
